Bugfix: silent failure of directory rename in extoolsExtract
Sometimes extoolsExtract would fail to rename the extracted directory to its final name. The error went unnoticed, so that extoolsLoad would fail to find it.

Use Files.move() instead of file.renameTo(), it throws an exception on failure, and the type of exception provides a hint of the root cause.